+++ src/renderer/components/attachment/messageAttachment.tsx
@@ -25,6 +25,7 @@ type AttachmentProps = {
   conversationType: 'group' | 'direct'
   direction: MessageType['msg']['direction']
   message: MessageType
+  hasQuote: boolean
 }
 
 export default function Attachment({
@@ -33,6 +34,7 @@ export default function Attachment({
   conversationType,
   direction,
   message,
+  hasQuote,
 }: AttachmentProps) {
   const tx = useTranslationFunction()
   if (!attachment) {
@@ -53,7 +55,7 @@ export default function Attachment({
   // For attachments which aren't full-frame
   const withContentBelow = withCaption
   const withContentAbove =
-    conversationType === 'group' && direction === 'incoming'
+    hasQuote || (conversationType === 'group' && direction === 'incoming')
   // const dimensions = message.msg.dimensions || {}
   // Calculating height to prevent reflow when image loads
   // const height = Math.max(MINIMUM_IMG_HEIGHT, (dimensions as any).height || 0)
@@ -134,6 +136,58 @@ export default function Attachment({
         )}
         onClick={onClickAttachment}
       >
+        <div
+          className='file-icon'
+          draggable='true'
+          onDragStart={dragAttachmentOut.bind(null, attachment)}
+          title={contentType}
+        >
+          {extension ? (
+            <div className='file-extension'>
+              {contentType === 'application/octet-stream' ? '' : extension}
+            </div>
+          ) : null}
+        </div>
+        <div className='text-part'>
+          <div className='name'>{fileName}</div>
+          <div className='size'>{fileSize}</div>
+        </div>
+      </div>
+    )
+  }
+}
+
+export function DraftAttachment({
+  attachment,
+}: {
+  attachment: MessageTypeAttachment
+}) {
+  if (!attachment) {
+    return null
+  }
+  if (isImage(attachment)) {
+    return (
+      <div className={classNames('message-attachment-media')}>
+        <img className='attachment-content' src={attachment.url} />
+      </div>
+    )
+  } else if (isVideo(attachment)) {
+    return (
+      <div className={classNames('message-attachment-media')}>
+        <video className='attachment-content' src={attachment.url} controls />
+      </div>
+    )
+  } else if (isAudio(attachment)) {
+    return (
+      <audio controls className={classNames('message-attachment-audio')}>
+        <source src={attachment.url} />
+      </audio>
+    )
+  } else {
+    const { fileName, fileSize, contentType } = attachment
+    const extension = getExtension(attachment)
+    return (
+      <div className={classNames('message-attachment-generic')}>
         <div
           className='file-icon'
           draggable='true'