+++ src/renderer/components/message/messageFunctions.ts
@@ -3,8 +3,9 @@ const { openItem } = window.electron_functions
 import { getLogger } from '../../../shared/logger'
 const log = getLogger('render/msgFunctions')
 import { Message } from 'deltachat-node'
-import { MessageType } from '../../../shared/shared-types'
-import { ChatStoreDispatch } from '../../stores/chat'
+import { MessageType, JsonMessage } from '../../../shared/shared-types'
+import { ChatStoreDispatch, selectChat } from '../../stores/chat'
+import { DeltaBackend } from '../../delta-remote'
 /**
  * json representation of the message object we get from the backend
  */
@@ -41,4 +42,30 @@ export function deleteMessage(
 
 export function openMessageInfo(message: MessageType) {
   window.__openDialog('MessageDetail', { id: message.id })
+}
+
+export function setQuoteInDraft(messageId: number) {
+  if (window.__setQuoteInDraft) {
+    window.__setQuoteInDraft(messageId)
+  } else {
+    throw new Error('window.__setQuoteInDraft undefined')
+  }
+}
+
+export async function privateReply(msg: JsonMessage) {
+  const quotedMessageId = msg.id
+  const contactId = msg.fromId
+  const chatId = await DeltaBackend.call('contacts.getDMChatId', contactId) // getDMChatId creates the dm chat if it doesn't exist
+
+  // retrieve existing draft to append the quotedMessageId
+  const oldDraft = await DeltaBackend.call('messageList.getDraft', chatId)
+
+  await DeltaBackend.call('messageList.setDraft', chatId, {
+    text: oldDraft?.msg.text,
+    file: oldDraft?.msg.file,
+    quotedMessageId,
+  })
+
+  // select chat
+  selectChat(chatId)
 }